# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Soccer",
        "description": "API Soccer.",
        "contact": {
            "name": "PT XYZ",
            "url": "https://xyz.test"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/players": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Get All Players",
                "operationId": "a355f9879160729f824bdea53e739b05",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Players retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "2"
                                                    },
                                                    "team_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Andi"
                                                    },
                                                    "height": {
                                                        "type": "integer",
                                                        "example": "175"
                                                    },
                                                    "weight": {
                                                        "type": "integer",
                                                        "example": "60"
                                                    },
                                                    "position": {
                                                        "type": "string",
                                                        "example": "Penyerang"
                                                    },
                                                    "jersey_number": {
                                                        "type": "integer",
                                                        "example": "11"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Team not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "404"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Team not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Store a new players",
                "operationId": "4e2752a0305de0adfa2a8c9ac1aa37a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Bekasi FC"
                                    },
                                    "height": {
                                        "type": "integer",
                                        "example": "175"
                                    },
                                    "weight": {
                                        "type": "integer",
                                        "example": "67"
                                    },
                                    "position": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "penyerang"
                                    },
                                    "jersey_number": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "team_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Player created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "one of fields must br required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statuscode": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}